requirements.txt
# CIC-IDS2017 Preprocessing Pipeline
# Install: pip install -r requirements.txt
# NOTE: Using >= pins for Python 3.14 compatibility (binary wheels).

# --- Core data processing ---
pandas>=2.2.3
numpy>=1.26.4
pyarrow>=18.1.0          # Fast Parquet / Feather I/O

# --- ML & preprocessing ---
scikit-learn>=1.6.1
imbalanced-learn>=0.12.4  # SMOTE, ADASYN, SMOTEENN, SMOTETomek
xgboost>=2.1.3

# --- Config & serialisation ---
PyYAML>=6.0.2
joblib>=1.4.2

# --- Visualisation ---
matplotlib>=3.10.0
seaborn>=0.13.2

# --- Utilities ---
tqdm>=4.67.1              # Progress bars for large file processing
